ADC refactor Peter Comments v2

+++ cpu/samd21/periph/uart.c
@@ -144,11 +144,11 @@ int uart_init_blocking(uart_t uart, uint32_t baudrate)
             /* Turn on power manager for sercom */
             PM->APBCMASK.reg |= PM_APBCMASK_SERCOM2;
 
-            /* configure GCLK0 to feed sercom0 */;
-            GCLK->CLKCTRL.reg = (uint32_t)((GCLK_CLKCTRL_CLKEN | GCLK_CLKCTRL_GEN_GCLK0 | (SERCOM2_GCLK_ID_CORE << GCLK_CLKCTRL_ID_Pos)));
+            /* configure GCLK0 to feed sercom2 */;
+            GCLK->CLKCTRL.reg = (uint16_t)((GCLK_CLKCTRL_CLKEN | GCLK_CLKCTRL_GEN_GCLK0 | (SERCOM2_GCLK_ID_CORE << GCLK_CLKCTRL_ID_Pos)));
             while (GCLK->STATUS.bit.SYNCBUSY);
 
-            GCLK->CLKCTRL.reg = (uint32_t)((GCLK_CLKCTRL_CLKEN | GCLK_CLKCTRL_GEN_GCLK0 | (SERCOM2_GCLK_ID_SLOW << GCLK_CLKCTRL_ID_Pos)));
+            GCLK->CLKCTRL.reg = (uint16_t)((GCLK_CLKCTRL_CLKEN | GCLK_CLKCTRL_GEN_GCLK0 | (SERCOM2_GCLK_ID_SLOW << GCLK_CLKCTRL_ID_Pos)));
             while (GCLK->STATUS.bit.SYNCBUSY);
 
             /* configure PINS to input/output*/
